page component move
