+++ src/pages/FAQ/FAQ.tsx
