+++ src/pages/FAQ/index.tsx
